+++ HISTORY.rst
@@ -25,6 +25,7 @@ latest (unreleased)
 
 * BSQOQ-34: Add controller for changing the shipping addresses of sale orders
 * BSQOQ-34: Add job to disable change of shipping address on sale orders
+* BSQOQ-40: Batch print of carrier delivery labels
 
 **Bugfixes**